# openapi/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.3
  title: Home Project Blog Service
  description: |
    Blog for the Home Project engineers
  license:
    name: MIT
    url: https://github.com/Home-Project-Engineering/home-project-blog/blob/main/LICENSE
servers:
  - url: http://localhost:8080/api/1
security:
  - basicAuth: []
...
paths:
...
  /tags:
    get:
      tags:
        - Tags
      summary: Get tags
      description: Uses query keys to retrieve a subset of existing tags.
      operationId: getTags
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/name'
        - $ref: '#/components/parameters/tags_sort'
        - $ref: '#/components/parameters/page_num'
        - $ref: '#/components/parameters/page_size'
      responses:
        "200":
          $ref: '#/components/responses/TagsResponse'
        "default":
          $ref: '#/components/responses/UnknownError'
  /tags/{id}:
    parameters:
      - $ref: '#/components/parameters/p_id'
    get:
      tags:
        - Tags
      summary: Get Tag by ID
      description: Retrieves an existing Tag associated with the specified ID.
      operationId: getTag
      responses:
        "200":
          $ref: '#/components/responses/TagResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "default":
          $ref: '#/components/responses/UnknownError'
    delete:
      tags:
        - Tags
      summary: Remove Tag
      description: Deletes a Tag associated with a specified ID.
      operationId: removeTag
      responses:
        "204":
          $ref: '#/components/responses/NoContentResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "default":
          $ref: '#/components/responses/UnknownError'
components:
  headers:
    X-Total-Count:
      description: The total results for the particular query.
      schema:
        type: number
        format: int64
  schemas:
...
    Tag:
      required:
        - name
      type: object
      properties:
        id:
          description: The ID of the Tag.
          type: number
          format: int64
          readOnly: true
        name:
          description: The name of the Tag.
          type: string
          example: Java8
...
    Error:
      type: object
      description: Representation of an Error that can appear using the application.
      properties:
        code:
          description: The code of an error that describes the Error.
          type: string
          readOnly: true
        message:
          description: The message of an error that describes the Error.
          type: string
          readOnly: true
  responses:
...
    NoContentResponse:
      description: The request was succesfully processed.
    TagResponse:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Tag"
...
    TagsResponse:
      description: OK
      headers:
        X-Total-Count:
          $ref: '#/components/headers/X-Total-Count'
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Tag"
...
    ValidationError:
      description: The payload contains an error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    UnknownError:
      description: The unknown error appeard. Check your payload or contact support.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  parameters:
    p_id:
      in: path
      name: id
      schema:
        type: number
        format: int64
      required: true
...
    page_num:
      in: query
      name: page_num
      schema:
        type: integer
    page_size:
      in: query
      name: page_size
      schema:
        type: integer
    id:
      in: query
      name: id
      schema:
        type: number
        format: int64
    name:
      in: query
      name: name
      schema:
        type: string
...
    tags_sort:
      in: query
      name: sort
      description: |
        In order to execute *asc*, you need to specify in the search *id* or *name* parameter.
        In order to complete the *desc* sorting must be specified in the query parameter *-id*
      schema:
        type: string
        enum: [ "id", "-id", "name", "-name" ]
        default: "name"
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic